update bc code to run on amr database and instructions on how to run the code
